delay registration for list builtin
Because we don't know if the disabled list should be disabled or not, we
delay the registration of the type inference for the `list` builtin
until we do know. In this case, it is triggered during compiler setup,
at the time when the typingctx is parameterized with the
`disable_typed_list` flag.
